all source and test files converted to typescript, jest and linting improvements, old packages including babel removed
