+++ index.ts
@@ -1,1 +1,7 @@
 export { default as Money } from './src/Money/Money';
+export { default as Currency } from './src/Currency/Currency';
+export { default as CurrencyStore } from './src/Currency/CurrencyStore';
+export { default as WealthError } from './src/errors/WealthError';
+export { default as CurrencyMismatchError } from './src/errors/CurrencyMismatchError';
+export { default as InvalidCurrencyError } from './src/errors/InvalidCurrencyError';
+export { default as WrongInputError } from './src/errors/WrongInputError';
