--- src/index.js
@@ -1,7 +1,0 @@
-export {default as Money} from './Money/Money';
-export {default as Currency} from './Currency/Currency';
-export {default as CurrencyStore} from './Currency/CurrencyStore';
-export {default as WealthError} from './errors/WealthError';
-export {default as CurrencyMismatchError} from './errors/CurrencyMismatchError';
-export {default as InvalidCurrencyError} from './errors/InvalidCurrencyError';
-export {default as WrongInputError} from './errors/WrongInputError';
